// src/interfaces/server.ts
enum Commands {
  Reg = 'reg',
  Update_winners = 'update_winners',
  Create_room = 'create_room',
  Update_room = 'update_room',
  Add_user_to_room = 'add_user_to_room',
  Create_game = 'create_game',
  Add_ships = 'add_ships',
  Start_game = 'start_game',
  Attack = 'attack',
  RandomAttack = 'randomAttack',
  Turn = 'turn',
  Finish = 'finish',
}

export { Commands };